Add a test-helpers file.
